hot_dog_back/requirements.txt
matplotlib
tensorboard
tqdm
torch
torchvision
pytorch-lightning
optuna
numpy
Pillow
plotly
scikit-learn

fastapi
uvicorn[standard]
python-multipart
